fix(memoized-task): now canceled tasks are automatically retried
A canceled task could occur if your component is torn down before the
task completes. For long running tasks (such as those network-bound),
this can be a big problem, because the task will become cached, and
the next time a component is mounted that wants to access that same
network resource, no remote request would occur because the canceled
task was cached.

Now, canceled tasks bust the cache, allowing quick navigations or
quick remounts to correctly re-fetch network resources (or whatever
the long running async task is)